+++ DynaLock/IMonitor.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace DynaLock
+{
+    public interface IMonitor : IDisposable
+    {
+        bool IsLockOwned();
+        void Enter();
+        bool TryEnter(int millisecondsTimeout = 0);
+    }
+}