+++ CHANGELOG.md
@@ -1,2 +1,7 @@
 # ChangeLog
 All notable changes to this project will be documented in this file.
+
+## 1.0.0 - 2020-08-27
+### Added
+- Initialized DynaLock
+- Added Monitor to the project